core/imagegen/stable_diffusion/requirements.txt
accelerate>=0.16.0
torchvision
transformers>=4.25.1
datasets
ftfy
tensorboard
Jinja2

# Optional
wandb
xformers
bitsandbytes